1038 improvements
Passing references to treenodes prevents the need to clone the Rc,
turning `x.as_ref.map(Rc::clone)` into just `&x`. This is my first time
working with the Rc<RefCell<_>> pattern so the initial attempt was just
fighting with the borrow checker.

Making `sum` a mutable reference removes the need to decide on how to
handle the recursion without double counting, each node now just adds
its value to the same accumulating reference.

There is now no need for a return, and so the None branch is no longer
needed either and the `match` becomes an `if let`

No need to borrow the node 3 times, factored out into a single
`borrow_mut`

The arithmetic
            let val = n.val;
            n.val += *sum;
            *sum += val;
calculates the same thing twice, so the temporary can be removed and the
second addition replaced with an assignment