1579 - Update UnionFind to use size union and path splitting
Union optimization:
Choosing to add smaller trees to bigger ones keeps the height
of the graph logarithmic in the worst case.  Over the test data
this doesn't seem to make a difference, but I may want to reuse
this implementation.

Find (path compaction) optimization:
Updating every node to its grandparents does twice as many
writes as are needed to reduce the height of any node by 1/2.
If a node's parent moves closer to the root, so does the node.
The loop is restructured to be more readible.